Add files via upload
This free virtual training hands-on experience is designed to give you broad exposure to the different functionalities of the Power BI product. 

At the end of the training, you'll be able to: 

• Connect to, import, and transform data from a variety of sources 

• Build real data models 
• Author and publish business intelligence reports 

You'll also learn how to customize and securely share your creations for collaboration with other groups inside your organization, including through mobile devices.